refactor(runtime): share tool result text extraction
